TASK: Apply migration Neos.Flow-20201205172733
This migration does not actually change any code. It just displays a
warning if a PHP file still refers to the no longer existing
ComponentInterface

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.